requirements.txt
google-generativeai
mem0ai
supabase
vecs
voyageai
google-auth-oauthlib
google-api-python-client
RealtimeSTT
openwakeword
piper-tts
pyaudio
rich
prompt_toolkit
schedule
python-dotenv
pyyaml
fastapi
uvicorn[standard]
python-multipart
